Align filename and definition in output
Instead of using tabs to align the three columns, use width so they're
all aligned. This makes reading the output much easier because you don't
have to parse for a colon.

Also, instead of showing the full filename, show just the tail. This
gives much more space for showing the definition and is likely the most
useful information. Otherwise the filename takes up all available space
and our columns never line up. (a:tag.filename is an absolute path for
me.)

I've tested the format_tag path, but I don't understand how to trigger
the read_line_fmt path. Maybe it's for old versions of vim?