When called with no args, use last jumped tag
Change `:Unity tselect` to behave like `:tselect`. This allows you to
use this command as a drop-in replacement for tselect:

  command! -nargs=* Tselect Unite tselect:<args>

The differences I'm aware of are unite's filtering and using cword
instead of an error message if tag stack is empty. And unite doesn't
push the current location onto the tagstack.